requirements.txt
asgiref==3.8.1
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-rest-knox==5.0.1
djangorestframework==3.15.2
dkimpy==1.1.8
dnspython==2.7.0
drf-yasg==1.21.7
dynaconf==3.2.6
environs==11.0.0
et-xmlfile==1.1.0
hvac==2.3.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
knox==0.1.14
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mysqlclient==2.2.4
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
PyPDF2==3.0.1
pyspf==2.0.14
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
setuptools==75.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
win32-setctime==1.1.0
XlsxWriter==3.2.0
dkimpy==1.1.8
dnspython==2.7.0
pyspf==2.0.14